requirements.txt
beautifulsoup4==4.8.0
blessings==1.7
certifi==2019.6.16
chardet==3.0.4
config42==0.3.2
decorator==4.0.11
fake-useragent==0.1.11
future==0.17.1
huepy==0.9.8.1
idna==2.8
imageio==2.1.2
instabot==0.55.0
instabot-py==0.4.10
InstagramAPI==1.0.2
instaloader==4.2.8
Jinja2==2.10.1
lxml==4.4.1
MarkupSafe==1.1.1
MechanicalSoup==0.11.0
moviepy==0.2.3.2
numpy==1.17.0
pandas==0.25.1
Pillow==6.1.0
pycollada==0.6
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1
requests==2.22.0
requests-toolbelt==0.9.1
responses==0.10.6
schedule==0.6.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.4
tqdm==4.35.0
urllib3==1.25.3
